feature/fix_today_bookmark_viewmodel_delete
