+++ app/src/main/java/com/millie/millieshelf/presentation/todayTop/viewHolder/TodayBookmarkViewHolder.kt
@@ -4,7 +4,6 @@ import androidx.recyclerview.widget.RecyclerView
 import com.millie.millieshelf.databinding.ItemTodayBookmarkBinding
 import com.millie.millieshelf.model.TodayTopModel
 import com.millie.millieshelf.presentation.todayTop.bookmark.TodayBookmarkAdapter
-import com.millie.millieshelf.presentation.todayTop.bookmark.TodayBookmarkViewModel
 
 class TodayBookmarkViewHolder(private val binding: ItemTodayBookmarkBinding) :
     RecyclerView.ViewHolder(binding.root) {
@@ -14,7 +13,7 @@ class TodayBookmarkViewHolder(private val binding: ItemTodayBookmarkBinding) :
 
         binding.rvTodayBookmark.apply {
             adapter = TodayBookmarkAdapter(context)
-            (adapter as TodayBookmarkAdapter).submitList(TodayBookmarkViewModel().mockTodayBookmarkList)
+            (adapter as TodayBookmarkAdapter).submitList(todayData.buttonList)
         }
     }
 }
